requirements.txt
aiogram==2.22.1
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.1
Django==4.1
django-ckeditor==6.5.0
django-cleanup==6.0.0
django-js-asset==2.0.0
django-modeltranslation==0.18.4
django-rosetta==0.9.8
frozenlist==1.3.1
googletrans==3.1.0a0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.8.1
html2text==2020.1.16
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
loguru==0.6.0
multidict==6.0.2
Pillow==9.2.0
polib==1.1.1
python-dotenv==0.20.0
pytils==0.4.1
pytz==2022.2.1
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
urllib3==1.26.12
yarl==1.8.1
